fix(committersStore): change test directory structure
