+++ tests/js/spec/stores/committers/withCommitters.spec.tsx
@@ -2,119 +2,9 @@ import {Fragment} from 'react';
 
 import {render, screen} from 'sentry-test/reactTestingLibrary';
 
-import {committersReducer} from 'sentry/stores/commiters/committersReducer';
+import {CommittersProvider} from 'sentry/stores/commiters/committersProvider';
+import {useCommitters} from 'sentry/stores/commiters/useCommitters';
 import {withCommitters} from 'sentry/stores/commiters/withCommitters';
-
-describe('CommitersReducer', () => {
-  it('marks committer loading and clears old state', () => {
-    const newState = committersReducer(
-      {},
-      {
-        type: 'start loading',
-        payload: {
-          organizationSlug: 'org',
-          projectSlug: 'project slug',
-          eventId: '0',
-        },
-      }
-    );
-
-    for (const key in newState) {
-      expect(newState[key].committers).toEqual([]);
-      expect(newState[key].committersLoading).toBeTruthy();
-      expect(newState[key].committersError).toBeFalsy();
-    }
-  });
-
-  it('marks committer error', () => {
-    const newState = committersReducer(
-      {},
-      {
-        type: 'set error',
-        payload: {
-          organizationSlug: 'org',
-          projectSlug: 'project slug',
-          eventId: '0',
-        },
-      }
-    );
-
-    for (const key in newState) {
-      expect(newState[key].committers).toEqual([]);
-      expect(newState[key].committersLoading).toBeFalsy();
-      expect(newState[key].committersError).toBeTruthy();
-    }
-  });
-
-  it('adds committer and nulls loading and error state', () => {
-    const committers = [TestStubs.CommitAuthor()];
-    const newState = committersReducer(
-      {},
-      {
-        type: 'add committers',
-        payload: {
-          organizationSlug: 'org',
-          projectSlug: 'project slug',
-          eventId: '0',
-          committers,
-        },
-      }
-    );
-
-    for (const key in newState) {
-      expect(newState[key].committers).toEqual(committers);
-      expect(newState[key].committersLoading).toBeFalsy();
-      expect(newState[key].committersError).toBeFalsy();
-    }
-  });
-});
-
-const TestProviderComponent = () => {
-  const [committers] = useCommitters();
-  const nodes = Object.keys(committers).flatMap(key => {
-    return committers[key].committers.map(c => c.author.name);
-  });
-
-  return nodes.length > 0 ? (
-    <Fragment>{nodes}</Fragment>
-  ) : (
-    <Fragment>empty component</Fragment>
-  );
-};
-
-describe('CommittersProvider', () => {
-  it('renders with no state by default', () => {
-    render(
-      <CommittersProvider>
-        <TestProviderComponent />
-      </CommittersProvider>
-    );
-
-    expect(screen.getByText(/empty component/)).toBeInTheDocument();
-  });
-
-  it('renders with initial state', () => {
-    render(
-      <CommittersProvider
-        initialState={{
-          'initial state': {
-            committers: [
-              TestStubs.Commit({
-                author: TestStubs.CommitAuthor({name: 'testing commit'}),
-              }),
-            ],
-            committersError: false,
-            committersLoading: false,
-          },
-        }}
-      >
-        <TestProviderComponent />
-      </CommittersProvider>
-    );
-
-    expect(screen.getByText(/testing commit/)).toBeInTheDocument();
-  });
-});
 
 const TestHOCComponent = withCommitters(({committers}) => {
   const [state] = useCommitters();
